// src/data/vocab.generated.ts
export type ReinoId = 'pueblo' | 'selva' | 'desierto' | 'montanas' | 'ciudad';
export type EntryKind = 'noun' | 'adjective' | 'verb' | 'preposition' | 'conjunction' | 'question' | 'adverb' | 'number' | 'phrase' | 'colombian';